requirements.txt
requests~=2.29.0
PyYAML~=6.0.1
setuptools~=69.0.2